+++ src/atm_phy_aes/tmx/mo_vdf_sfc.f90
@@ -200,7 +200,8 @@ CONTAINS
       & albvisdir_tile, albvisdif_tile, &
       & albnirdir_tile, albnirdif_tile, &
       & kh_tile, km_tile, &
-      & kh_neutral_tile, km_neutral_tile, &
+      & interp_fac_2m_tile, interp_fac_10m_tile, &
+      & interp_fac_tsfc_tile, &
       & wind_rel_tile, &
       & lhfl_tile, shfl_tile, &
       & ustress_tile, vstress_tile, &
@@ -306,8 +307,9 @@ CONTAINS
     albedo_tile => diags%albedo_tile%Get_ptr_r3d()
     kh_tile => diags%kh_tile%Get_ptr_r3d()
     km_tile => diags%km_tile%Get_ptr_r3d()
-    kh_neutral_tile => diags%kh_neutral_tile%Get_ptr_r3d()
-    km_neutral_tile => diags%km_neutral_tile%Get_ptr_r3d()
+    interp_fac_2m_tile => diags%interp_fac_2m_tile%Get_ptr_r3d()
+    interp_fac_10m_tile => diags%interp_fac_10m_tile%Get_ptr_r3d()
+    interp_fac_tsfc_tile => diags%interp_fac_tsfc_tile%Get_ptr_r3d()
     lhfl_tile => diags%lhfl_tile%Get_ptr_r3d()
     shfl_tile => diags%shfl_tile%Get_ptr_r3d()
     wind_rel_tile => diags%wind_rel_tile%Get_ptr_r3d()
@@ -396,6 +398,18 @@ CONTAINS
         CALL init(albnirdir_tile(:,:,jtile), albedoW, lacc=.TRUE., opt_acc_async_queue=acc_async_queues(jtile))
         CALL init(albnirdif_tile(:,:,jtile), albedoW, lacc=.TRUE., opt_acc_async_queue=acc_async_queues(jtile))
 !$OMP END PARALLEL
+
+!$OMP PARALLEL DO PRIVATE(jc, jcl, jb) ICON_OMP_DEFAULT_SCHEDULE
+        DO jb = this%domain%i_startblk_c, this%domain%i_endblk_c
+          !$ACC PARALLEL LOOP DEFAULT(PRESENT) GANG VECTOR PRIVATE(jc) ASYNC(acc_async_queues(jtile))
+          DO jcl = 1, nvalid(jb,jtile)
+            jc = indices(jcl,jb,jtile)
+            interp_fac_tsfc_tile(jc,jb,jtile) = interp_fac_2m_tile(jc,jb,jtile) * new_tsfc(jc,jb,jtile)
+          END DO
+          !$ACC END PARALLEL LOOP
+        END DO
+!$OMP END PARALLEL DO
+
       CASE(isfc_ice)
         IF (nice_thickness_classes /= 1) CALL finish(routine, 'Only one ice thickness class (kice) implemented!')
 !$OMP PARALLEL
@@ -426,6 +440,7 @@ CONTAINS
             new_tsfc_rad(jc,jb,jtile) = new_tsfc(jc,jb,jtile)
             new_tsfc_eff(jc,jb,jtile) = new_tsfc(jc,jb,jtile)
             tend_tsfc(jc,jb,jtile) = (new_tsfc(jc,jb,jtile) - old_tsfc(jc,jb,jtile)) / dtime
+            interp_fac_tsfc_tile(jc,jb,jtile) = interp_fac_2m_tile(jc,jb,jtile) * new_tsfc(jc,jb,jtile)
           END DO
           !$ACC END PARALLEL LOOP
         END DO
@@ -453,8 +468,8 @@ CONTAINS
           & albvisdir_tile(:,:,jtile), albvisdif_tile(:,:,jtile), &
           & albnirdir_tile(:,:,jtile), albnirdif_tile(:,:,jtile), &
           & kh_tile(:,:,jtile), km_tile(:,:,jtile), &
-          & kh_neutral_tile(:,:,jtile), km_neutral_tile(:,:,jtile), &
-          & co2flx_nat_tile(:,:,jtile) &
+          & interp_fac_2m_tile(:,:,jtile), interp_fac_10m_tile(:,:,jtile), &
+          & interp_fac_tsfc_tile(:,:,jtile), co2flx_nat_tile(:,:,jtile) &
           & )
 
 !$OMP PARALLEL DO PRIVATE(jc, jcl, jb) ICON_OMP_DEFAULT_SCHEDULE
@@ -634,7 +649,7 @@ CONTAINS
     REAL(wp), POINTER, DIMENSION(:,:,:) :: &
       & qsat_tile, rho_tile, theta_tile, thetav_tile, &
       & moist_rich_tile, rough_h_tile, rough_m_tile, &
-      & km_tile, kh_tile, km_neutral_tile, kh_neutral_tile, &
+      & km_tile, kh_tile, interp_fac_2m_tile, interp_fac_10m_tile, &
       & evapotrans_tile, lhfl_tile, shfl_tile, &
       & wind_rel_tile, ustress_tile, vstress_tile, &
       & u10m_tile, v10m_tile, wind10m_tile, &
@@ -721,8 +736,8 @@ CONTAINS
     rough_m_tile => diags%rough_m_tile%Get_ptr_r3d()
     km_tile => diags%km_tile%Get_ptr_r3d()
     kh_tile => diags%kh_tile%Get_ptr_r3d()
-    km_neutral_tile => diags%km_neutral_tile%Get_ptr_r3d()
-    kh_neutral_tile => diags%kh_neutral_tile%Get_ptr_r3d()
+    interp_fac_2m_tile => diags%interp_fac_2m_tile%Get_ptr_r3d()
+    interp_fac_10m_tile => diags%interp_fac_10m_tile%Get_ptr_r3d()
     evapotrans_tile => diags%evapotrans_tile%Get_ptr_r3d()
     lhfl_tile => diags%lhfl_tile%Get_ptr_r3d()
     shfl_tile => diags%shfl_tile%Get_ptr_r3d()
@@ -825,12 +840,11 @@ CONTAINS
             & theta_tile(:,:,jtile), qsat_tile(:,:,jtile), &
             ! Output
             & km_tile(:,:,jtile), kh_tile(:,:,jtile), &
-            & km_neutral_tile(:,:,jtile), kh_neutral_tile(:,:,jtile))
+            & interp_fac_2m_tile(:,:,jtile), interp_fac_10m_tile(:,:,jtile))
           CALL compute_10m_wind( &
             & this%domain, nvalid(:,jtile), indices(:,:,jtile), &
-            & zf, zh, &
             & ua, va, &
-            & moist_rich_tile(:,:,jtile), km_tile(:,:,jtile), km_neutral_tile(:,:,jtile), &
+            & interp_fac_10m_tile(:,:,jtile), &
             & u10m_tile(:,:,jtile), v10m_tile(:,:,jtile), wind10m_tile(:,:,jtile) &
             & )
 
@@ -888,7 +902,7 @@ CONTAINS
           & theta_tile(:,:,jtile), qsat_tile(:,:,jtile), &
             ! Output
           & km_tile(:,:,jtile), kh_tile(:,:,jtile), &
-          & km_neutral_tile(:,:,jtile), kh_neutral_tile(:,:,jtile), &
+          & interp_fac_2m_tile(:,:,jtile), interp_fac_10m_tile(:,:,jtile), &
           & opt_acc_async_queue=jtile)
       END IF