+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
 - Simplification of the rte-rrtmgp radiation interface (https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/1087)
 - Fix loop indexing in TMX surface module (https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/1241)
 - Added options to not use Louis stability formula over land and/or sea ice
+- Fix in TMX of the near-surface diagnostics interpolation effecting 2m temperature and 10m wind
 
 #### NWP Physics